added animations and utils submodules to tools, cleaned up model plot functions
