+++ radd/models.py
@@ -14,7 +14,7 @@ class Simulator(object):
     """ Core code for simulating models. All cond, trials, &
     timepoints are simulated simultaneously
     """
-    def __init__(self, fitparams=None, pc_map=None, kind='xdpm', dt=.002, si=.01, learn=False):
+    def __init__(self, fitparams=None, pc_map=None, kind='xdpm', dt=.005, si=.01, learn=False):
         self.learn = learn
         self.kind = kind
         self.pc_map = pc_map
