+++ radd/tools/utils.py
@@ -1,6 +1,7 @@
 #!/usr/local/bin/env python
 from __future__ import division
 import sys
+from future.utils import listvalues
 from copy import deepcopy
 import pandas as pd
 import numpy as np
@@ -108,8 +109,6 @@ class NestedProgress(object):
                 bar.update(new_progress)
             elif bar.bartype=='colorbar':
                 bar.update(i=i, new_progress=new_progress)
-            else:
-                continue
 
     def callback(self, x, fmin, accept):
         """ A callback function for reporting basinhopping status
@@ -131,101 +130,81 @@ class NestedProgress(object):
         for bar in listvalues(self.bars):
             bar.clear()
 
-def render_animation(anim):
-    plt.close(anim._fig)
-    return HTML(anim_to_html(anim))
 
-def animate_dpm(model):
-    """ to render animation within a notebook :
-        vis.render_animation(vis.animated_dpm_example(MODEL))
+def rwr(X, get_index=False, n=None):
     """
-    from matplotlib import animation
-    params = deepcopy(model.inits)
-    bound=theta.scalarize_params(params)['a']
-    x, gtraces, straces, xi, yi, nframes = gen_re_traces(model, params)
-    f, axes = build_decision_axis(onset=x[0][0], bound=bound)
-    glines = [axes[i].plot([], [], linewidth=1.5)[0] for i, n in enumerate(gtraces)]
-    slines = [axes[i].plot([xi[i]], [yi[i]], linewidth=1.5)[0] for i, n in enumerate(straces)]
-    f_args = (x, gtraces, glines, straces, slines, params, xi, yi)
-    anim=animation.FuncAnimation(f, re_animate_multiax, fargs=f_args, frames=nframes, interval=1, blit=True)
-    return anim
+    Modified from http://nbviewer.ipython.org/gist/aflaxman/6871948
+    """
+    if isinstance(X, pd.Series):
+        X = X.copy()
+        X.index = range(len(X.index))
+    if n == None:
+        n = len(X)
+    resample_i = np.floor(np.random.rand(n) * len(X)).astype(int)
+    X_resample = (X[resample_i])
+    if get_index:
+        return resample_i
+    else:
+        return X_resample
 
-def gen_re_traces(model, params):
-    sim = deepcopy(model.opt.simulator)
-    sim.__update_steps__(dt=.001)
-    ssd, nssd, nss, nss_per, ssd_ix = sim.ssd_info
-    nsstot = nss * nssd
-    params = sim.vectorize_params(params)
-    bound = params['a'][0]
-    Pg, Tg, xtb = sim.__update_go_process__(params)
-    Ps, Ts = sim.__update_stop_process__(params)
-    Ts=[Ts[0, i] for i in [2, -1]]
-    dvg, dvs = sim.sim_fx(params, analyze=False)
-    dvg = dvg[0, :nss, :].reshape(nssd, nss_per, dvg.shape[-1])
-    gtraces = [dvg[i, 0] for i in [2, -1]]
-    straces = [dvs[0, i, 0] for i in [2, -1]]
-    gtraces = [gt[gt <= bound] for gt in gtraces]
-    ssi, xinit_ss, integrated, dvgs, dvss = [], [], [], [], []
-    for i, (g, s) in enumerate(zip(gtraces, straces)):
-        xinit_ss.append(Tg[0] - Ts[i])
-        ssi.append(g[:xinit_ss[i]])
-        ss = s[:Ts[i]]
-        s = np.append(g[:xinit_ss[i]], ss[ss >= 0])
-        ixmin = np.min([len(g), len(s)])
-        dvgs.append(g[:ixmin])
-        dvss.append(s[:ixmin])
-    nframes = [len(gt) for gt in dvgs]
-    x = params['tr'][0] * 1000 + [np.arange(nf) for nf in nframes]
-    sim.__update_steps__(dt=.005)
-    return [x, dvgs, dvss, xinit_ss, ssi, np.max(nframes)]
+def resample_data(data, n=120, groups=['ssd']):
+    """ generates n resampled datasets using rwr()
+    for bootstrapping model fits
+    """
+    df = data.copy()
+    bootlist = list()
+    if n == None:
+        n = len(df)
+    for level, level_df in df.groupby(groups):
+        boots = level_df.reset_index(drop=True)
+        orig_ix = np.asarray(boots.index[:])
+        resampled_ix = rwr(orig_ix, get_index=True, n=n)
+        bootdf = level_df.irow(resampled_ix)
+        bootlist.append(bootdf)
+    # concatenate and return all resampled conditions
+    return self.model.rangl_data(pd.concat(bootlist))
 
-def build_decision_axis(onset, bound, ssd=np.array([300, 400]), tb=650):
-    # init figure, axes, properties
-    f, axes = plt.subplots(len(ssd), 1, figsize=(7, 7), dpi=300)
-    #f.subplots_adjust(wspace=.05, top=.1, bottom=.1)
-    f.subplots_adjust(hspace=.05, top=.99, bottom=.09)
-    w = tb + 40
-    h = bound
-    start = onset - 80
-    for i, ax in enumerate(axes):
-        plt.setp(ax, xlim=(start - 1, w + 1), ylim=(0 - (.01 * h), h + (.01 * h)))
-        ax.vlines(x=ssd[i], ymin=0, ymax=h, color="#e74c3c", lw=2.5, alpha=.5)
-        ax.hlines(y=h, xmin=start, xmax=w, color='k')
-        ax.hlines(y=0, xmin=start, xmax=w, color='k')
-        ax.vlines(x=tb, ymin=0, ymax=h, color='#2043B0', lw=2.5, linestyle='-', alpha=.5)
-        ax.vlines(x=start + 2, ymin=0, ymax=h, color='k')
-        ax.text(ssd[i] + 10, h * .88, str(ssd[i]) + 'ms', fontsize=19)
-        ax.set_xticklabels([])
-        ax.set_yticklabels([])
-        ax.set_xticks([])
-        ax.set_yticks([])
-    sns.despine(top=True, right=True, bottom=True, left=True)
-    return f, axes
+def extract_popt_fitinfo(finfo=None, plist=None, pc_map=None):
+    """ takes optimized dict or DF of vectorized parameters and
+    returns dict with only depends_on.keys() containing vectorized vals.
+    Is accessed by fit.Optimizer objects after optimization routine.
+    ::Arguments::
+    finfo (dict/DF):
+        finfo is dict if self.fit_on is 'average'
+        and DF if self.fit_on is 'subjects' or 'bootstrap'
+        contains optimized parameters
+    ::Returns::
+    popt (dict):
+        dict with only depends_on.keys() containing
+        vectorized vals
+    """
+    finfo = dict(deepcopy(finfo))
+    plist = list(inits)
+    popt = {pkey: finfo[pkey] for pkey in plist}
+    for pkey in list(pc_map):
+        popt[pkey] = np.array([finfo[pc] for pc in pc_map[pkey]])
+    return popt
 
-def re_animate_multiax(i, x, gtraces, glines, straces, slines, params, xi, yi):
-    gcolor = '#27ae60'
-    scolor = '#e74c3c'
-    for n in range(len(x)):
-        ex, gt, gl, st, sl, ix, iy = [xx[n] for xx in [x, gtraces, glines, straces, slines, xi, yi]]
-        try:
-            gl.set_data(ex[:i + 1], gt[:i + 1])
-            gl.set_color(gcolor)
-            sl.set_data(ex[ix:i + 1], st[ix:i + 1])
-            sl.set_color(scolor)
-        except Exception:
-            return sl, gl
-        #f.savefig('animation_frames/movie/img' + str(i) +'.png', dpi=300)
-    return sl, gl
+def params_io(p={}, io='w', iostr='popt'):
+    """ read // write parameters dictionaries
+    """
+    if io == 'w':
+        pd.Series(p).to_csv(''.join([iostr, '.csv']))
+    elif io == 'r':
+        ps = pd.read_csv(''.join([iostr, '.csv']), header=None)
+        p = dict(zip(ps[0], ps[1]))
+        return p
 
-def anim_to_html(anim):
-    from tempfile import NamedTemporaryFile
-    VIDEO_TAG = """<video controls>
-             <source src="data:video/x-m4v;base64,{0}" type="video/mp4">
-             Your browser does not support the video tag.
-            </video>"""
-    if not hasattr(anim, '_encoded_video'):
-        with NamedTemporaryFile(suffix='.mp4') as f:
-            anim.save(f.name, extra_args=['-vcodec', 'libx264', '-pix_fmt', 'yuv420p'])
-            video = open(f.name, "rb").read()
-        anim._encoded_video = video.encode("base64")
-    return VIDEO_TAG.format(anim._encoded_video)
+def fits_io(fitparams, fits=[], io='w', iostr='fits'):
+    """ read // write y, wts, yhat arrays
+    """
+    y = fitparams['y'].flatten()
+    wts = fitparams['wts'].flatten()
+    fits = fits.flatten()
+    if io == 'w':
+        index = np.arange(len(fits))
+        df = pd.DataFrame({'y': y, 'wts': wts, 'yhat': fits}, index=index)
+        df.to_csv(''.join([iostr, '.csv']))
+    elif io == 'r':
+        df = pd.read_csv(''.join([iostr, '.csv']), index_col=0)
+        return df
