+++ radd/vis.py
@@ -36,81 +36,70 @@ def unpack_vector(vector, nlevels=1, nquant=5, nssd=1, kde_quant=False):
         quant_cor = quant[:nquant]
         quant_err = quant[-nquant:]
         if kde_quant:
-            quant_cor = analyze.kde_fit_quantiles([quant_cor], bw=.005)
-            quant_err = analyze.kde_fit_quantiles([quant_err], bw=.005)
+            quant_cor = analyze.kde_fit_quantiles([quant_cor], bw=.01)
+            quant_err = analyze.kde_fit_quantiles([quant_err], bw=.01)
         unpacked.append([presp, quant_cor, quant_err])
     return unpacked
 
 def plot_model_fits(y, yhat, fitparams, kde_quant=True, palette=bpal, save=False, cdf=True):
+    """ main plotting function for displaying model fit predictions over data
+    """
     nlevels = y.ndim
     fitparams['nlevels'] = nlevels
     quantiles = fitparams['quantiles']
     ssd, nssd, nss, nss_per, ssd_ix = fitparams['ssd_info']
-    f, axes = plt.subplots(nlevels, 3, figsize=(12, 3.5*nlevels), sharey=True)
+    f, axes = plt.subplots(1, 3, figsize=(14, 4), sharey=True)
     axes = np.asarray(axes)
     clrs_lvl = palette(nlevels)
     y_data = unpack_vector(y, nlevels=nlevels, nquant=quantiles.size, nssd=nssd, kde_quant=False)
-    yhat_data = unpack_vector(yhat, nlevels=nlevels, nquant=quantiles.size, nssd=nssd, kde_quant=True)
+    yhat_data = unpack_vector(yhat, nlevels=nlevels, nquant=quantiles.size, nssd=nssd, kde_quant=False)
+    y_kde = unpack_vector(y, nlevels=nlevels, nquant=quantiles.size, nssd=nssd, kde_quant=True)
+    yhat_kde = unpack_vector(yhat, nlevels=nlevels, nquant=quantiles.size, nssd=nssd, kde_quant=True)
     if nssd>1:
         ssds = ssd.mean(axis=0).squeeze()*1e3
         scx = np.array([np.float(d) for d in ssds])
-        plot_func = scurves
+        plot_acc = scurves
     else:
         scx = np.array([0, 1])
-        plot_func = plot_accuracy
+        plot_acc = plot_accuracy
     labels = get_plot_labels(fitparams)
-
     for i in range(nlevels):
-        if nlevels>1:
-            ax1, ax2, ax3 = axes[i]
-        else:
-            ax1, ax2, ax3 = axes
-        stop_acc = [y_data[i][0], yhat_data[i][0]]
-        qc, qc_hat = [y_data[i][1], yhat_data[i][1]]
-        qe, qe_hat = [y_data[i][2], yhat_data[i][2]]
+        ax1, ax2, ax3 = axes
         c = clrs_lvl[i]
         lemp, lhat = labels[i]
-        plot_func(stop_acc, x=scx, colors=[c, c], markers=False, labels=labels[i], linestyles=['-', '--'], ax=ax1)
-        sns.kdeplot(qc_hat, cumulative=cdf, color=c, ax=ax2, linewidth=2, linestyle='--', bw=.01)
-        sns.kdeplot(qe_hat, cumulative=cdf, color=c, ax=ax3, linewidth=2, linestyle='--', bw=.01)
-        ax2.plot(qc, quantiles, color=c, linewidth=0, marker='o')
-        ax3.plot(qe, quantiles, color=c, linewidth=0, marker='o')
+        plot_acc([y_data[i][0], yhat_data[i][0]], x=scx, colors=[c,c], labels=labels[i], ax=ax1)
+        plot_quantiles(y_data[i], yhat_data[i], color=c, axes=[ax2, ax3], kde=False, labels=labels[i])
+        plot_quantiles(y_kde[i], yhat_kde[i], color=c, axes=[ax2, ax3], kde=True)
 
-        if i==0:
-            ax1.set_title('Stop Accuracy', position=(.5, 1.02))
-            ax2.set_title('Correct RT Quantiles', position=(.5, 1.02))
-            ax3.set_title('Error RT Quantiles', position=(.5, 1.02))
-            if nlevels>1:
-                for ax in axes[0, :]:
-                    ax.set_xticklabels([])
-    for ax in axes[:, -2:].flatten():
+    ax1.set_title('Stop Accuracy', position=(.5, 1.02))
+    ax2.set_title('Correct RT Quantiles', position=(.5, 1.02))
+    ax3.set_title('Error RT Quantiles', position=(.5, 1.02))
+    for ax in [ax2, ax3]:
         qdata = np.hstack([arr[1:] for arr in yhat_data]).flatten()
-        xxlim = (qdata.min()-.03, qdata.max()+.03)
-        ax.set_xlim(xxlim[0], xxlim[1])
-    for ax in axes.flatten()[-2:]:
-        ax.set_xlabel('RT (s)')
+        xxlim = (qdata.min()-.05, qdata.max()+.05)
         xxticks = [xxlim[0], np.mean(xxlim), xxlim[-1]]
         xxtls = ["{:.2f}".format(xtl) for xtl in xxticks]
-        plt.setp(ax, xlim=xxlim, xticks=xxticks, xticklabels=xxtls)
-    axes[-1, 0].set_xlabel('SSD (ms)')
+        plt.setp(ax, xlim=xxlim, xticks=xxticks, xticklabels=xxtls, xlabel='RT (s)')
+    ax1.set_xlabel('SSD (ms)')
+    plt.tight_layout()
     if save:
         savestr = '_'.join([fitparams['kind'], str(fitparams['idx'])+'.png'])
         plt.savefig(savestr, dpi=600)
         plt.close('all')
 
-def scurves(lines=[],  yerr=[], pretune_fx='interpolate', ax=None, linestyles=None, colors=None, labels=None, mc=None, x=None, get_pse=False, **kwargs):
-
+def scurves(lines=[],  ax=None, colors=None, labels=None, x=None, get_pse=False, pretune_fx='interpolate'):
+    """ plotting function for displaying model-predicted
+    stop curve (across SSDs) over empirical estimates
+    """
     if ax is None:
         f, ax = plt.subplots(1, figsize=(5.5, 6))
     if colors is None:
         colors = slate(len(lines))
     if labels is None:
         labels = [''] * len(lines)
-    if linestyles is None:
-        linestyles = ['-'] * len(lines)
-    lines = [(line) if type(line) == list else line for line in lines]
     if x is None:
         x = np.arange(len(lines[0]), dtype='float')*50
+    fc = (0.,0.,0.,.01)
     pse=[]
     for i, yi in enumerate(lines):
         if pretune_fx=='interpolate':
@@ -121,15 +110,12 @@ def scurves(lines=[],  yerr=[], pretune_fx='interpolate', ax=None, linestyles=No
         idx = (np.abs(ysim - .5)).argmin()
         pse.append(xsim[idx])
         # Plot the results
-        if yerr != []:
-            ax.errorbar(x, yi, yerr=yerr[i], color=colors[i], lw=0, marker='o', elinewidth=2, ecolor=colors[i])
-            ax.plot(xsim, ysim, linestyle=linestyles[i], lw=2., color=colors[i], label=labels[i])
-        else:
-            if not i%2:
-                ax.plot(x, yi, lw=0., color=colors[i], label=labels[i], marker='o')
-            else:
-                ax.plot(xsim, ysim, linestyle=linestyles[i], lw=2.1, color=colors[i], label=labels[i])
-        pse.append(xsim[idx])
+        if i%2:
+            ax.plot(x, yi, lw=0., color=colors[i],  marker='o', ms=5)
+            ax.plot(xsim, ysim, lw=2, color=colors[i], label=labels[i])
+            continue
+        ax.plot(x, yi, lw=0., color=colors[i], marker='o', ms=9, mfc=fc, mec=colors[i], mew=1.5)
+        ax.plot(xsim, ysim, linestyle='--', lw=2, color=colors[i], label=labels[i])
     xxlim = (xsim[0], xsim[-1])
     plt.setp(ax, xticks=x, xlim=xxlim, ylim=(-.01, 1.05), yticks=np.arange(0, 1.2, .2), ylabel='P(Stop)')
     ax.legend(loc=0)
@@ -139,7 +125,34 @@ def scurves(lines=[],  yerr=[], pretune_fx='interpolate', ax=None, linestyles=No
     if get_pse:
         return pse
 
+def plot_quantiles(y_data, yhat_data, axes=None, color=None, labels=None, kde=False, quantiles=np.array([.1, .3, .5, .7, .9]), bw=.01):
+    """ plotting function for displaying model-predicted
+    quantile-probabilities over empirical estimates
+    """
+    fc = (0.,0.,0.,.01)
+    c = color
+    if axes is not None:
+        ax2, ax3 = axes
+    else:
+        f, (ax2, ax3) = plt.subplots(1, 2, figsize=(10, 4))
+    qc, qc_hat = [y_data[1], yhat_data[1]]
+    qe, qe_hat = [y_data[2], yhat_data[2]]
+    if kde:
+        sns.kdeplot(qc, cumulative=1, color=c, ax=ax2, linewidth=2, linestyle='-', bw=bw)
+        sns.kdeplot(qe, cumulative=1, color=c, ax=ax3, linewidth=2, linestyle='-', bw=bw)
+        sns.kdeplot(qc_hat, cumulative=1, color=c, ax=ax2, linewidth=2, linestyle='--', bw=bw)
+        sns.kdeplot(qe_hat, cumulative=1, color=c, ax=ax3, linewidth=2, linestyle='--', bw=bw)
+    else:
+        ax2.plot(qc, quantiles, color=c, linewidth=0, marker='o', ms=5, label=labels)
+        ax3.plot(qe, quantiles, color=c, linewidth=0, marker='o', ms=5, label=labels)
+        ax2.plot(qc_hat, quantiles, mec=c, linewidth=0, marker='o', ms=8, mfc=fc, mew=1.5, label=labels)
+        ax3.plot(qe_hat, quantiles, mec=c, linewidth=0, marker='o', ms=8, mfc=fc, mew=1.5, label=labels)
+
 def plot_accuracy(lines=[], yerr=None, x=np.array([0, 1]), ax=None, linestyles=None, colors=None, labels=None, **kwargs):
+    """ plotting function for displaying model-predicted
+    stop probability (at mean SSD) on top of empirical estimates
+    (used when data is collected using tracking procedure to est. SSRT)
+    """
     if ax is None:
         f, ax = plt.subplots(1, figsize=(5.5, 6))
     if colors is None:
